+++ sesdev.py
@@ -69,8 +69,12 @@ def _print_log(output):
               help='Libvirt storage pool')
 @click.option('--stop-before-deepsea-stage', type=int, default=None,
               help='Allows to stop deployment before running the specified DeepSea stage')
+@click.option('--deployment-tool', type=str, default=None,
+              help='Deployment tool to deploy the Ceph cluster. Currently only deepsea is supported')
+@click.option('--version', type=click.Choice(['ses5', 'ses6', 'ses7', 'luminous', 'nautilus', 'octopus']), default=None,
+              help='SES version to install (ses5, ses6, luminous, nautilus, octopus)')
 def create(deployment_id, roles, os, deploy, num_disks, single_node, deepsea_cli, libvirt_host,
-           libvirt_user, libvirt_storage_pool, stop_before_deepsea_stage):
+           libvirt_user, libvirt_storage_pool, stop_before_deepsea_stage, deployment_tool, version):
     settings_dict = {}
     if not single_node and roles:
         roles = [r.strip() for r in roles.split(",")]
@@ -120,6 +124,12 @@ def create(deployment_id, roles, os, deploy, num_disks, single_node, deepsea_cli
 
     if stop_before_deepsea_stage is not None:
         settings_dict['stop_before_stage'] = stop_before_deepsea_stage
+
+    if deployment_tool is not None:
+        settings_dict['deployment_tool'] = deployment_tool
+
+    if version is not None:
+        settings_dict['version'] = version
 
     settings = seslib.Settings(**settings_dict)